refactor(components): PR 评审面板 diff/comments/drafts 收归 features/pr/
diff / comments / drafts 仅在 PR 详情视图内渲染，属 PR 作用域子领域，迁入
features/pr/ 下，使 pr/ 成为「PR 评审工作区」的统一归属（与 PrItem /
PrInfoView / CommitsPanel 同处）。chat / onboarding / settings 仍在 features/
顶层。纯目录迁移 + import 路径改写（git 识别为 rename），无逻辑/界面变更。
